+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 - Add changelog to docs and hex info. Thanks @epinault!
+- Add `Moar.Term.blank?/2` and `Moar.Term.present?/2` which can override what values are considered blank & present.
 
 ## 4.1.0